Move user resources to dedicated pkg
